decouple posting list from visitor using the reader abstraction
