+++ lib/segment/src/index/field_index/full_text_index/compressed_common.rs
@@ -1,0 +1,1 @@
+pub type BitPackerImpl = bitpacking::BitPacker4x;
