+++ lib/segment/src/index/field_index/full_text_index/mod.rs
@@ -5,5 +5,8 @@ mod postings_iterator;
 pub mod text_index;
 mod tokenizers;
 
+mod compressed_chunks_reader;
+pub mod compressed_common;
+mod compressed_posting_visitor;
 #[cfg(test)]
 mod tests;
